+++ apps/decodex/src/agent/json_rpc.rs
@@ -1,765 +1,21 @@
-use std::{
-	collections::VecDeque,
-	env,
-	ffi::OsString,
-	fmt::{self, Display, Formatter},
-	io::{self, BufRead as _, BufReader, Write as _},
-	path::{Path, PathBuf},
-	process::{Child, ChildStdin, Command, Stdio},
-	sync::{
-		Arc, Mutex,
-		mpsc::{self, Receiver, RecvTimeoutError},
-	},
-	thread,
-	time::Duration,
+mod connection;
+mod environment;
+mod errors;
+mod wire;
+
+#[cfg(test)] pub(crate) use self::wire::JsonRpcErrorPayload;
+pub(crate) use self::{
+	connection::JsonRpcConnection,
+	environment::{AppServerProcessEnv, ResolvedAppServerCodexHomeEnv, app_server_command_program},
+	errors::{AppServerHomePreflightFailure, AppServerOutputTimeout, AppServerTransportFailure},
+	wire::{JsonRpcError, JsonRpcMessage, JsonRpcNotification, JsonRpcRequest, WireMessage},
 };
 
-use color_eyre::{Report, eyre};
-use serde::{Deserialize, Serialize, de::DeserializeOwned};
-use serde_json::{self, Value};
-
-use crate::git_credentials::{GitCredentialEnvironment, GitSigningConfig};
-
-const APP_SERVER_STDERR_TAIL_LINES: usize = 20;
-const CODEX_APP_SERVER_BINARY: &str = "codex";
-const CODEX_HOME_ENV_VAR: &str = "CODEX_HOME";
-const CODEX_SQLITE_HOME_ENV_VAR: &str = "CODEX_SQLITE_HOME";
-const CODEX_HOME_DIR_NAME: &str = ".codex";
-
-#[derive(Clone, Eq, PartialEq)]
-pub(crate) struct AppServerProcessEnv {
-	git: GitCredentialEnvironment,
-	codex_home_policy: AppServerCodexHomePolicy,
-}
-impl AppServerProcessEnv {
-	#[cfg(test)]
-	pub(crate) fn with_github_credentials(
-		github_token_env_var: String,
-		github_token: String,
-	) -> Self {
-		Self {
-			git: GitCredentialEnvironment::with_github_credentials(
-				github_token_env_var,
-				github_token,
-			),
-			codex_home_policy: AppServerCodexHomePolicy::SharedDefault,
-		}
-	}
-
-	pub(crate) fn with_github_credentials_and_signing_config(
-		github_token_env_var: String,
-		github_token: String,
-		signing_config: GitSigningConfig,
-	) -> Self {
-		Self {
-			git: GitCredentialEnvironment::with_github_credentials_and_signing_config(
-				github_token_env_var,
-				github_token,
-				signing_config,
-			),
-			codex_home_policy: AppServerCodexHomePolicy::SharedDefault,
-		}
-	}
-
-	pub(crate) fn resolve_codex_home_env(
-		&self,
-	) -> crate::prelude::Result<ResolvedAppServerCodexHomeEnv> {
-		match &self.codex_home_policy {
-			AppServerCodexHomePolicy::SharedDefault => resolve_shared_codex_home_env(),
-			#[cfg(test)]
-			AppServerCodexHomePolicy::Explicit(home_env) => Ok(home_env.clone()),
-		}
-	}
-
-	pub(crate) fn apply_to(
-		&self,
-		command: &mut Command,
-	) -> crate::prelude::Result<ResolvedAppServerCodexHomeEnv> {
-		self.git.apply_to(command);
-
-		let codex_home_env = self.resolve_codex_home_env()?;
-
-		codex_home_env.apply_to(command)?;
-
-		Ok(codex_home_env)
-	}
-
-	#[cfg(test)]
-	fn with_codex_home_for_test(home_env: ResolvedAppServerCodexHomeEnv) -> Self {
-		Self {
-			git: GitCredentialEnvironment::default(),
-			codex_home_policy: AppServerCodexHomePolicy::Explicit(home_env),
-		}
-	}
-}
-
-impl Default for AppServerProcessEnv {
-	fn default() -> Self {
-		Self {
-			git: GitCredentialEnvironment::default(),
-			codex_home_policy: AppServerCodexHomePolicy::SharedDefault,
-		}
-	}
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct ResolvedAppServerCodexHomeEnv {
-	codex_home: PathBuf,
-	sqlite_home: PathBuf,
-}
-impl ResolvedAppServerCodexHomeEnv {
-	pub(crate) fn new(codex_home: PathBuf, sqlite_home: PathBuf) -> crate::prelude::Result<Self> {
-		validate_codex_home_path(CODEX_HOME_ENV_VAR, &codex_home)?;
-		validate_codex_home_path(CODEX_SQLITE_HOME_ENV_VAR, &sqlite_home)?;
-
-		Ok(Self { codex_home, sqlite_home })
-	}
-
-	pub(crate) fn codex_home(&self) -> &Path {
-		&self.codex_home
-	}
-
-	#[cfg(test)]
-	fn sqlite_home(&self) -> &Path {
-		&self.sqlite_home
-	}
-
-	fn apply_to(&self, command: &mut Command) -> crate::prelude::Result<()> {
-		let codex_home = path_env_value(CODEX_HOME_ENV_VAR, &self.codex_home)?;
-		let sqlite_home = path_env_value(CODEX_SQLITE_HOME_ENV_VAR, &self.sqlite_home)?;
-
-		command.env_remove(CODEX_HOME_ENV_VAR);
-		command.env_remove(CODEX_SQLITE_HOME_ENV_VAR);
-		command.env(CODEX_HOME_ENV_VAR, codex_home);
-		command.env(CODEX_SQLITE_HOME_ENV_VAR, sqlite_home);
-
-		Ok(())
-	}
-}
-
-#[derive(Debug)]
-pub(crate) struct AppServerHomePreflightFailure {
-	details: String,
-	kind: AppServerHomePreflightFailureKind,
-}
-impl AppServerHomePreflightFailure {
-	pub(crate) fn resolution_failed(details: String) -> Self {
-		Self { details, kind: AppServerHomePreflightFailureKind::ResolutionFailed }
-	}
-
-	pub(crate) fn initialize_mismatch(resolved_home: String, expected_home: String) -> Self {
-		Self {
-			details: format!(
-				"app_server_protocol_failure: initialize codexHome `{resolved_home}` did not match expected shared Codex home `{expected_home}`; Decodex blocked dispatch before thread/start so Codex state is not split across homes."
-			),
-			kind: AppServerHomePreflightFailureKind::InitializeMismatch,
-		}
-	}
-
-	pub(crate) fn error_class(&self) -> &'static str {
-		match self.kind {
-			AppServerHomePreflightFailureKind::ResolutionFailed => {
-				"app_server_codex_home_preflight_failed"
-			},
-			AppServerHomePreflightFailureKind::InitializeMismatch => {
-				"app_server_codex_home_mismatch"
-			},
-		}
-	}
-
-	pub(crate) fn terminal_next_action(&self, recovery_gate: &str) -> String {
-		format!(
-			"inspect the local Decodex and Codex home sharing, restart `decodex serve`, {recovery_gate}"
-		)
-	}
-}
-
-impl Display for AppServerHomePreflightFailure {
-	fn fmt(&self, formatter: &mut Formatter<'_>) -> fmt::Result {
-		formatter.write_str(&self.details)
-	}
-}
-
-impl std::error::Error for AppServerHomePreflightFailure {}
-
-#[derive(Debug)]
-pub(crate) struct AppServerOutputTimeout;
-impl Display for AppServerOutputTimeout {
-	fn fmt(&self, formatter: &mut Formatter<'_>) -> fmt::Result {
-		formatter.write_str("Timed out while waiting for app-server output.")
-	}
-}
-
-impl std::error::Error for AppServerOutputTimeout {}
-
-#[derive(Debug)]
-pub(crate) struct AppServerTransportFailure {
-	details: String,
-	phase: Option<&'static str>,
-	retryable_startup: bool,
-}
-impl AppServerTransportFailure {
-	pub(crate) fn new(details: String) -> Self {
-		Self { details, phase: None, retryable_startup: false }
-	}
-
-	pub(crate) fn with_phase(
-		details: String,
-		phase: &'static str,
-		retryable_startup: bool,
-	) -> Self {
-		Self { details, phase: Some(phase), retryable_startup }
-	}
-
-	pub(crate) fn error_class(&self) -> &'static str {
-		"app_server_transport_disconnected"
-	}
-
-	pub(crate) fn is_retryable_startup(&self) -> bool {
-		self.retryable_startup
-	}
-
-	pub(crate) fn retry_next_action(&self) -> String {
-		if let Some(phase) = self.phase {
-			format!(
-				"app-server transport disconnected during `{phase}` before a durable turn was running; decodex will restart the app-server and retry automatically"
-			)
-		} else {
-			String::from("decodex will retry the app-server transport failure automatically")
-		}
-	}
-
-	pub(crate) fn terminal_next_action(&self, recovery_gate: &str) -> String {
-		let phase = self.phase.map_or_else(String::new, |phase| format!(" during `{phase}`"));
-
-		format!(
-			"inspect the local app-server stderr tail and process exit status, resolve the Codex app-server transport failure{phase} manually, {recovery_gate}"
-		)
-	}
-}
-
-impl Display for AppServerTransportFailure {
-	fn fmt(&self, formatter: &mut Formatter<'_>) -> fmt::Result {
-		formatter.write_str(&self.details)
-	}
-}
-
-impl std::error::Error for AppServerTransportFailure {}
-
-pub(crate) struct JsonRpcConnection {
-	child: Child,
-	stdin: ChildStdin,
-	stdout_rx: Receiver<String>,
-	stderr_tail: Arc<Mutex<VecDeque<String>>>,
-	pending_messages: VecDeque<WireMessage>,
-	next_request_id: i64,
-}
-impl JsonRpcConnection {
-	pub(crate) fn spawn_app_server(
-		listen: &str,
-		process_env: &AppServerProcessEnv,
-	) -> crate::prelude::Result<Self> {
-		let mut command = Command::new(app_server_command_program());
-		let _codex_home_env = configure_app_server_command(&mut command, listen, process_env)?;
-		let mut child = command.spawn()?;
-		let stdin =
-			child.stdin.take().ok_or_else(|| eyre::eyre!("Failed to capture app-server stdin."))?;
-		let stdout = child
-			.stdout
-			.take()
-			.ok_or_else(|| eyre::eyre!("Failed to capture app-server stdout."))?;
-		let stderr = child
-			.stderr
-			.take()
-			.ok_or_else(|| eyre::eyre!("Failed to capture app-server stderr."))?;
-		let (stdout_tx, stdout_rx) = mpsc::channel();
-		let stderr_tail = Arc::new(Mutex::new(VecDeque::new()));
-		let _stdout_task = thread::spawn(move || {
-			let reader = BufReader::new(stdout);
-
-			for line in reader.lines() {
-				match line {
-					Ok(line) => {
-						let line: String = line;
-
-						if line.trim().is_empty() {
-							continue;
-						}
-						if stdout_tx.send(line).is_err() {
-							break;
-						}
-					},
-					Err(error) => {
-						tracing::warn!(?error, "Failed to read app-server stdout.");
-
-						break;
-					},
-				}
-			}
-		});
-		let stderr_tail_writer = Arc::clone(&stderr_tail);
-		let _stderr_task = thread::spawn(move || {
-			let reader = BufReader::new(stderr);
-
-			for line in reader.lines() {
-				match line {
-					Ok(line) => {
-						let line: String = line;
-						let trimmed_line = line.trim().to_owned();
-
-						if trimmed_line.is_empty() {
-							continue;
-						}
-
-						match stderr_tail_writer.lock() {
-							Ok(mut tail) => {
-								if tail.len() == APP_SERVER_STDERR_TAIL_LINES {
-									tail.pop_front();
-								}
-
-								tail.push_back(trimmed_line);
-							},
-							Err(error) => {
-								tracing::warn!(?error, "Failed to retain app-server stderr tail.");
-							},
-						}
-
-						tracing::warn!(stderr = %line, "codex app-server stderr");
-					},
-					Err(error) => {
-						tracing::warn!(?error, "Failed to read app-server stderr.");
-
-						break;
-					},
-				}
-			}
-		});
-
-		Ok(Self {
-			child,
-			stdin,
-			stdout_rx,
-			stderr_tail,
-			pending_messages: VecDeque::new(),
-			next_request_id: 1,
-		})
-	}
-
-	#[allow(dead_code)]
-	pub(crate) fn request<P, T>(
-		&mut self,
-		method: &str,
-		params: &P,
-		timeout: Duration,
-	) -> crate::prelude::Result<T>
-	where
-		P: Serialize,
-		T: DeserializeOwned,
-	{
-		self.request_with_handler(method, params, timeout, |_connection, _message, request| {
-			eyre::bail!(
-				"Unexpected inbound JSON-RPC request `{}` while waiting for `{method}`.",
-				request.method
-			);
-		})
-	}
-
-	pub(crate) fn request_with_handler<P, T, F>(
-		&mut self,
-		method: &str,
-		params: &P,
-		timeout: Duration,
-		mut handle_request: F,
-	) -> crate::prelude::Result<T>
-	where
-		P: Serialize,
-		T: DeserializeOwned,
-		F: FnMut(&mut Self, &WireMessage, &JsonRpcRequest) -> crate::prelude::Result<()>,
-	{
-		let request_id = self.next_request_id;
-		let expected_id = Value::from(request_id);
-
-		self.next_request_id += 1;
-
-		self.send_value(&serde_json::json!({
-			"id": request_id,
-			"method": method,
-			"params": params,
-		}))?;
-
-		loop {
-			let wire_message = self.read_message(Some(timeout))?;
-
-			match &wire_message.message {
-				JsonRpcMessage::Notification(_) => self.pending_messages.push_back(wire_message),
-				JsonRpcMessage::Response(response) if response.id == expected_id => {
-					return Ok(serde_json::from_value(response.result.clone())?);
-				},
-				JsonRpcMessage::Error(error) if error.id == expected_id => {
-					let data = error
-						.error
-						.data
-						.as_ref()
-						.map_or_else(String::new, |data| format!(" data: {data}"));
-
-					return Err(eyre::eyre!(
-						"`{method}` failed with {}: {}{}",
-						error.error.code,
-						error.error.message,
-						data
-					));
-				},
-				JsonRpcMessage::Request(request) => handle_request(self, &wire_message, request)?,
-				JsonRpcMessage::Response(response) => {
-					tracing::debug!(
-						method,
-						response_id = %response.id,
-						expected_id = %expected_id,
-						"Recorded and ignored orphan app-server JSON-RPC response while waiting for request."
-					);
-				},
-				JsonRpcMessage::Error(error) => {
-					return Err(eyre::eyre!(
-						"Received an unexpected JSON-RPC error while waiting for `{method}`: id {} failed with {}: {}",
-						error.id,
-						error.error.code,
-						error.error.message
-					));
-				},
-			}
-		}
-	}
-
-	pub(crate) fn notify<P>(
-		&mut self,
-		method: &str,
-		params: Option<&P>,
-	) -> crate::prelude::Result<()>
-	where
-		P: Serialize,
-	{
-		let value = match params {
-			Some(params) => serde_json::json!({
-				"method": method,
-				"params": params,
-			}),
-			None => serde_json::json!({ "method": method }),
-		};
-
-		self.send_value(&value)
-	}
-
-	pub(crate) fn recv(
-		&mut self,
-		timeout: Option<Duration>,
-	) -> crate::prelude::Result<WireMessage> {
-		if let Some(message) = self.pending_messages.pop_front() {
-			return Ok(message);
-		}
-
-		self.read_message(timeout)
-	}
-
-	pub(crate) fn respond<R>(&mut self, id: &Value, result: &R) -> crate::prelude::Result<()>
-	where
-		R: Serialize,
-	{
-		self.send_value(&serde_json::json!({
-			"id": id,
-			"result": result,
-		}))
-	}
-
-	pub(crate) fn respond_error(
-		&mut self,
-		id: &Value,
-		code: i64,
-		message: &str,
-	) -> crate::prelude::Result<()> {
-		self.send_value(&serde_json::json!({
-			"id": id,
-			"error": {
-				"code": code,
-				"message": message,
-			},
-		}))
-	}
-
-	pub(crate) fn drain_pending(&mut self) -> Vec<WireMessage> {
-		self.pending_messages.drain(..).collect()
-	}
-
-	fn send_value(&mut self, value: &Value) -> crate::prelude::Result<()> {
-		let payload = serde_json::to_string(value)?;
-
-		if let Err(error) = writeln!(self.stdin, "{payload}") {
-			return Err(self.app_server_stdin_error("write", error));
-		}
-		if let Err(error) = self.stdin.flush() {
-			return Err(self.app_server_stdin_error("flush", error));
-		}
-
-		Ok(())
-	}
-
-	fn read_message(&mut self, timeout: Option<Duration>) -> crate::prelude::Result<WireMessage> {
-		let raw = match timeout {
-			Some(timeout) => match self.stdout_rx.recv_timeout(timeout) {
-				Ok(raw) => raw,
-				Err(RecvTimeoutError::Timeout) => {
-					return Err(Report::new(AppServerOutputTimeout));
-				},
-				Err(RecvTimeoutError::Disconnected) => {
-					return Err(self.app_server_disconnect_error());
-				},
-			},
-			None => self.stdout_rx.recv().map_err(|_| self.app_server_disconnect_error())?,
-		};
-
-		WireMessage::parse(raw)
-	}
-
-	fn app_server_disconnect_error(&mut self) -> Report {
-		let details = self.app_server_transport_error_details(
-			"App-server stdout disconnected unexpectedly".to_owned(),
-		);
-
-		Report::new(AppServerTransportFailure::new(details))
-	}
-
-	fn app_server_stdin_error(&mut self, operation: &str, error: io::Error) -> Report {
-		let details = self.app_server_transport_error_details(format!(
-			"App-server stdin {operation} failed: {error}"
-		));
-
-		Report::new(AppServerTransportFailure::new(details))
-	}
-
-	fn app_server_transport_error_details(&mut self, summary: String) -> String {
-		let process_status = match self.child.try_wait() {
-			Ok(Some(status)) => format!("process exited with `{status}`"),
-			Ok(None) => String::from("process was still running"),
-			Err(error) => format!("failed to inspect process status: {error}"),
-		};
-		let stderr_tail = self.stderr_tail_snapshot();
-		let mut details = format!("{summary} ({process_status}).");
-
-		if !stderr_tail.is_empty() {
-			details.push_str(" Recent app-server stderr tail:");
-
-			for line in stderr_tail {
-				details.push_str("\n  ");
-				details.push_str(&line);
-			}
-		}
-
-		details
-	}
-
-	fn stderr_tail_snapshot(&self) -> Vec<String> {
-		match self.stderr_tail.lock() {
-			Ok(tail) => tail.iter().cloned().collect(),
-			Err(error) => {
-				tracing::warn!(?error, "Failed to read app-server stderr tail.");
-
-				Vec::new()
-			},
-		}
-	}
-}
-
-impl Drop for JsonRpcConnection {
-	fn drop(&mut self) {
-		let _ = self.child.kill();
-		let _ = self.child.wait();
-	}
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct WireMessage {
-	pub(crate) raw: String,
-	pub(crate) message: JsonRpcMessage,
-}
-impl WireMessage {
-	fn parse(raw: String) -> crate::prelude::Result<Self> {
-		let value: Value = serde_json::from_str(&raw)?;
-		let message = if value.get("method").is_some() && value.get("id").is_some() {
-			JsonRpcMessage::Request(serde_json::from_value(value)?)
-		} else if value.get("method").is_some() {
-			JsonRpcMessage::Notification(serde_json::from_value(value)?)
-		} else if value.get("error").is_some() {
-			JsonRpcMessage::Error(serde_json::from_value(value)?)
-		} else if value.get("result").is_some() {
-			JsonRpcMessage::Response(serde_json::from_value(value)?)
-		} else {
-			return Err(eyre::eyre!("Received an unrecognized JSON-RPC payload: {raw}"));
-		};
-
-		Ok(Self { raw, message })
-	}
-}
-
-#[derive(Clone, Debug, Deserialize)]
-pub(crate) struct JsonRpcRequest {
-	pub(crate) id: Value,
-	pub(crate) method: String,
-	#[serde(default)]
-	pub(crate) params: Value,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-pub(crate) struct JsonRpcNotification {
-	pub(crate) method: String,
-	#[serde(default)]
-	pub(crate) params: Value,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-pub(crate) struct JsonRpcResponse {
-	pub(crate) id: Value,
-	pub(crate) result: Value,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-pub(crate) struct JsonRpcError {
-	pub(crate) id: Value,
-	pub(crate) error: JsonRpcErrorPayload,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-pub(crate) struct JsonRpcErrorPayload {
-	pub(crate) code: i64,
-	pub(crate) message: String,
-
-	pub(crate) data: Option<Value>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) enum JsonRpcMessage {
-	Request(JsonRpcRequest),
-	Notification(JsonRpcNotification),
-	Response(JsonRpcResponse),
-	Error(JsonRpcError),
-}
-
-#[derive(Debug)]
-enum AppServerHomePreflightFailureKind {
-	ResolutionFailed,
-	InitializeMismatch,
-}
-
-#[derive(Clone, Eq, PartialEq)]
-enum AppServerCodexHomePolicy {
-	SharedDefault,
-	#[cfg(test)]
-	Explicit(ResolvedAppServerCodexHomeEnv),
-}
-
-pub(crate) fn app_server_command_program() -> PathBuf {
-	app_server_command_program_from_env(env::var_os("PATH"), env::var_os("HOME"))
-}
-
-fn resolve_shared_codex_home_env() -> crate::prelude::Result<ResolvedAppServerCodexHomeEnv> {
-	resolve_shared_codex_home_env_from_home(env::var_os("HOME"))
-}
-
-fn resolve_shared_codex_home_env_from_home(
-	home: Option<OsString>,
-) -> crate::prelude::Result<ResolvedAppServerCodexHomeEnv> {
-	let Some(home) = home else {
-		return Err(Report::new(AppServerHomePreflightFailure::resolution_failed(String::from(
-			"app_server_preflight_failed: HOME is not set, so Decodex cannot resolve the shared Codex home for app-server dispatch.",
-		))));
-	};
-	let home = PathBuf::from(home);
-
-	if home.as_os_str().is_empty() {
-		return Err(Report::new(AppServerHomePreflightFailure::resolution_failed(String::from(
-			"app_server_preflight_failed: HOME is empty, so Decodex cannot resolve the shared Codex home for app-server dispatch.",
-		))));
-	}
-	if !home.is_absolute() {
-		return Err(Report::new(AppServerHomePreflightFailure::resolution_failed(format!(
-			"app_server_preflight_failed: HOME `{}` is not absolute, so Decodex cannot resolve the shared Codex home for app-server dispatch.",
-			home.display()
-		))));
-	}
-
-	let codex_home = home.join(CODEX_HOME_DIR_NAME);
-
-	ResolvedAppServerCodexHomeEnv::new(codex_home.clone(), codex_home)
-}
-
-fn validate_codex_home_path(name: &str, path: &Path) -> crate::prelude::Result<()> {
-	if path.as_os_str().is_empty() {
-		return Err(Report::new(AppServerHomePreflightFailure::resolution_failed(format!(
-			"app_server_preflight_failed: {name} resolved to an empty path."
-		))));
-	}
-	if !path.is_absolute() {
-		return Err(Report::new(AppServerHomePreflightFailure::resolution_failed(format!(
-			"app_server_preflight_failed: {name} `{}` is not absolute.",
-			path.display()
-		))));
-	}
-
-	path_env_value(name, path).map(|_| ())
-}
-
-fn path_env_value(name: &str, path: &Path) -> crate::prelude::Result<String> {
-	path.to_str().map(str::to_owned).ok_or_else(|| {
-		Report::new(AppServerHomePreflightFailure::resolution_failed(format!(
-			"app_server_preflight_failed: {name} `{}` is not valid UTF-8.",
-			path.display()
-		)))
-	})
-}
-
-fn configure_app_server_command(
-	command: &mut Command,
-	listen: &str,
-	process_env: &AppServerProcessEnv,
-) -> crate::prelude::Result<ResolvedAppServerCodexHomeEnv> {
-	command
-		.args(["app-server", "--listen", listen])
-		.stdin(Stdio::piped())
-		.stdout(Stdio::piped())
-		.stderr(Stdio::piped());
-
-	process_env.apply_to(command)
-}
-
-fn app_server_command_program_from_env(
-	path_env: Option<OsString>,
-	home: Option<OsString>,
-) -> PathBuf {
-	if let Some(path_env) = path_env {
-		for path_entry in env::split_paths(&path_env) {
-			let candidate = path_entry.join(CODEX_APP_SERVER_BINARY);
-
-			if candidate.is_file() {
-				return candidate;
-			}
-		}
-	}
-	if let Some(home) = home {
-		let home = PathBuf::from(home);
-
-		for relative_candidate in
-			[[".local", "bin", CODEX_APP_SERVER_BINARY], [".cargo", "bin", CODEX_APP_SERVER_BINARY]]
-		{
-			let candidate = relative_candidate
-				.iter()
-				.fold(home.clone(), |path, component| path.join(*component));
-
-			if candidate.is_file() {
-				return candidate;
-			}
-		}
-	}
-
-	PathBuf::from(CODEX_APP_SERVER_BINARY)
-}
+#[cfg(test)]
+use self::environment::{
+	app_server_command_program_from_env, configure_app_server_command,
+	resolve_shared_codex_home_env_from_home,
+};
 
 #[cfg(test)]
 mod tests {
